+++ source/newtonian/two_dimensional/source_terms/ConservativeForce.cpp
@@ -15,29 +15,29 @@ namespace {
 		Edge edge;
 		for(int i=0;i<n;++i)
 		{
-			edge=tess.GetEdge(edge_index[i]);
+		  edge=tess.GetEdge(edge_index[(size_t)i]);
 			if(point==edge.neighbors.first)
 				if(edge.neighbors.second>-1)
 				{
-					dm-=lengthes[edge_index[i]]*fluxes[edge_index[i]].Mass*
+				  dm-=lengthes[(size_t)edge_index[(size_t)i]]*fluxes[(size_t)edge_index[(size_t)i]].Mass*
 						(center-tess.GetMeshPoint(edge.neighbors.second));
 				}
 				else
 				{
 					Vector2D r=hbc.Normal(edge,tess);
-					dm-=2*lengthes[edge_index[i]]*fluxes[edge_index[i]].Mass*r;
+					dm-=2*lengthes[(size_t)edge_index[(size_t)i]]*fluxes[(size_t)edge_index[(size_t)i]].Mass*r;
 				}
 			else
 				if(point==edge.neighbors.second)
 					if(edge.neighbors.first>-1)
 					{
-						dm+=lengthes[edge_index[i]]*fluxes[edge_index[i]].Mass*
+					  dm+=lengthes[(size_t)edge_index[(size_t)i]]*fluxes[(size_t)edge_index[(size_t)i]].Mass*
 							(center-tess.GetMeshPoint(edge.neighbors.first));
 					}
 					else
 					{
 						Vector2D r=hbc.Normal(edge,tess);
-						dm+=2*lengthes[edge_index[i]]*fluxes[edge_index[i]].Mass*r;
+						dm+=2*lengthes[(size_t)edge_index[(size_t)i]]*fluxes[(size_t)edge_index[(size_t)i]].Mass*r;
 					}
 				else
 					throw UniversalError("Error in ConservativeForce MassFlux: Cell and edge are not mutual neighbors");
@@ -70,8 +70,8 @@ Conserved ConservativeForce::Calculate
 		 tracer_extensive,
 		 time, dt);
 	double volume=tess.GetVolume(point);
-	res.Momentum=volume*cells[point].Density*acc;
-	res.Energy=cells[point].Density*volume*ScalarProd(point_velocity[point],acc)+
+	res.Momentum=volume*cells[(size_t)point].Density*acc;
+	res.Energy=cells[(size_t)point].Density*volume*ScalarProd(point_velocity[(size_t)point],acc)+
 		0.5*ScalarProd(MassFlux(tess,point,fluxes,hbc,lengthes),acc);
 	if(DtCalc_)
 	{
